+++ src/controller/Error.js
@@ -14,32 +14,32 @@ const ERROR = Object.freeze({
  */
 const ERROR_HANDLING = Object.freeze({
     "[ERROR] 3이상 20이하 사이의 숫자여야 합니다.\n": (bridgeGame, message) => {
-        const ErrorView = require('./view/ErrorView');
-        const InputView = require('./view/InputView');
+        const ErrorView = require('../view/ErrorView');
+        const InputView = require('../view/InputView');
         ErrorView.printError(message);
         InputView.readBridgeSize(bridgeGame);
     },
     "[ERROR] 입력값은 숫자여야 합니다.\n": (bridgeGame, message) => {
-        const ErrorView = require('./view/ErrorView');
-        const InputView = require('./view/InputView');
+        const ErrorView = require('../view/ErrorView');
+        const InputView = require('../view/InputView');
         ErrorView.printError(message);
         InputView.readBridgeSize(bridgeGame);
     },
     "[ERROR] 입력값은 U 또는 D 여야합니다.\n": (bridgeGame, message) => {
-        const ErrorView = require('./view/ErrorView');
-        const InputView = require('./view/InputView');
+        const ErrorView = require('../view/ErrorView');
+        const InputView = require('../view/InputView');
         ErrorView.printError(message);
         InputView.readMoving(bridgeGame);
     },
     '[ERROR] 건널 수 없는 곳 입니다.\n': (bridgeGame) => {
-        const OutputView = require('./view/OutputView');
-        const InputView = require('./view/InputView');
+        const OutputView = require('../view/OutputView');
+        const InputView = require('../view/InputView');
         OutputView.printMap(bridgeGame);
         InputView.readGameCommand(bridgeGame);
     },
     '[ERROR] 입력값은 R 또는 Q 여야합니다.\n': (bridgeGame, message) => {
-        const ErrorView = require('./view/ErrorView');
-        const InputView = require('./view/InputView');
+        const ErrorView = require('../view/ErrorView');
+        const InputView = require('../view/InputView');
         ErrorView.printError(message);
         InputView.readGameCommand(bridgeGame);
     }
